Moved some items to "src" directory
