--- src/shared/App.tsx
@@ -1,15 +1,0 @@
-import {BrowserRouter} from "react-router-dom";
-import Routes from "./components/Routes.tsx";
-
-function App() {
-
-  return (
-    <>
-        <BrowserRouter>
-            <Routes />
-        </BrowserRouter>
-    </>
-  )
-}
-
-export default App
